refactor(core): split embed extension into focused modules
embed.ts mixed four concerns in 1231 lines: the provider catalogue and
oEmbed fetch logic, the NodeView DOM renderers, the Markdown metadata
hooks, and the node spec itself. Extract the first three into focused
co-located modules.

- utils/embed-providers.ts: pure, SSR-safe provider detection and oEmbed
  fetch with no DOM, now testable in isolation.
- extensions/embed-renderers.ts: NodeView DOM renderers and the Twitter
  script loader.
- extensions/embed-markdown.ts: metadata-comment encoding and the
  markdown-it inline rule.

embed.ts keeps the node spec, options, and view-model helpers, and
re-exports the moved public symbols so the consumer import surface is
unchanged. link-editor.ts imports detectVizelEmbedProvider from utils/,
removing a builder-to-extension dependency.
