+++ packages/core/src/extensions/embed.ts
@@ -11,349 +11,31 @@
 import { mergeAttributes, Node } from "@tiptap/core";
 import type { Node as PMNode } from "@tiptap/pm/model";
 import { Plugin, PluginKey } from "@tiptap/pm/state";
-import DOMPurify from "dompurify";
 import type MarkdownIt from "markdown-it";
 import type { MarkdownSerializerState } from "prosemirror-markdown";
+import {
+  createVizelDefaultFetchEmbedData,
+  detectVizelEmbedProvider,
+  type VizelEmbedData,
+  type VizelEmbedProvider,
+  type VizelEmbedType,
+  type VizelFetchEmbedDataFn,
+  vizelDefaultEmbedProviders,
+} from "../utils/embed-providers.ts";
 import { emitVizelError, VizelError } from "../utils/errorHandling.ts";
+import { escapeMetadataCommentValue, registerEmbedRule } from "./embed-markdown.ts";
+import { renderLink, renderLoading, renderOembed, renderOgpCard } from "./embed-renderers.ts";
 
-/**
- * Embed type based on available metadata
- */
-export type VizelEmbedType = "oembed" | "ogp" | "title" | "link";
-
-/**
- * oEmbed provider definition
- */
-export interface VizelEmbedProvider {
-  /** Provider name (e.g., 'youtube', 'twitter') */
-  name: string;
-  /** URL patterns to match */
-  patterns: RegExp[];
-  /** oEmbed API endpoint (optional - some providers don't support oEmbed) */
-  oEmbedEndpoint?: string;
-  /** Whether the oEmbed endpoint supports CORS (can be fetched from browser) */
-  supportsCors?: boolean;
-  /** Transform function for URL (e.g., extract video ID) */
-  transform?: (url: string) => string;
-}
-
-/**
- * Embed metadata returned from fetch
- */
-export interface VizelEmbedData {
-  /** Original URL */
-  url: string;
-  /** Type of embed based on available data */
-  type: VizelEmbedType;
-  /** Provider name if detected */
-  provider?: string;
-  /** oEmbed HTML content */
-  html?: string;
-  /** oEmbed width */
-  width?: number;
-  /** oEmbed height */
-  height?: number;
-  /** OGP/oEmbed title */
-  title?: string;
-  /** OGP/oEmbed description */
-  description?: string;
-  /** OGP/oEmbed image URL */
-  image?: string;
-  /** OGP site name */
-  siteName?: string;
-  /** Favicon URL */
-  favicon?: string;
-  /** Whether the embed is currently loading */
-  loading?: boolean;
-}
-
-/**
- * Function to fetch embed data from URL
- * This should be provided by the user to handle CORS
- */
-export type VizelFetchEmbedDataFn = (url: string) => Promise<VizelEmbedData>;
-
-/**
- * Default oEmbed providers.
- *
- * The exported array is readonly so consumers cannot mutate the shared
- * defaults. To extend the list, spread it into a new array.
- */
-export const vizelDefaultEmbedProviders: readonly VizelEmbedProvider[] = [
-  // Video - these providers support CORS on their oEmbed endpoints
-  {
-    name: "youtube",
-    patterns: [
-      /^https?:\/\/(www\.)?youtube\.com\/watch\?v=[\w-]+/,
-      /^https?:\/\/youtu\.be\/[\w-]+/,
-      /^https?:\/\/(www\.)?youtube\.com\/embed\/[\w-]+/,
-      /^https?:\/\/(www\.)?youtube\.com\/shorts\/[\w-]+/,
-    ],
-    oEmbedEndpoint: "https://www.youtube.com/oembed",
-    supportsCors: true,
-  },
-  {
-    name: "vimeo",
-    patterns: [/^https?:\/\/(www\.)?vimeo\.com\/\d+/],
-    oEmbedEndpoint: "https://vimeo.com/api/oembed.json",
-    supportsCors: true,
-  },
-  {
-    name: "loom",
-    patterns: [/^https?:\/\/(www\.)?loom\.com\/share\/[\w-]+/],
-    oEmbedEndpoint: "https://www.loom.com/v1/oembed",
-    supportsCors: true,
-  },
-  {
-    name: "tiktok",
-    patterns: [/^https?:\/\/(www\.)?tiktok\.com\/@[\w.-]+\/video\/\d+/],
-    oEmbedEndpoint: "https://www.tiktok.com/oembed",
-    supportsCors: true,
-  },
-  // Social - these do NOT support CORS, require server-side proxy
-  {
-    name: "twitter",
-    patterns: [
-      /^https?:\/\/(twitter\.com|x\.com)\/\w+\/status\/\d+/,
-      /^https?:\/\/(twitter\.com|x\.com)\/\w+\/moments\/\d+/,
-    ],
-    oEmbedEndpoint: "https://publish.twitter.com/oembed",
-    supportsCors: false,
-  },
-  {
-    name: "instagram",
-    patterns: [
-      /^https?:\/\/(www\.)?instagram\.com\/(p|reel)\/[\w-]+/,
-      /^https?:\/\/instagr\.am\/(p|reel)\/[\w-]+/,
-    ],
-    oEmbedEndpoint: "https://graph.facebook.com/v10.0/instagram_oembed",
-    supportsCors: false,
-  },
-  // Dev - CodePen and CodeSandbox support CORS
-  {
-    name: "codepen",
-    patterns: [/^https?:\/\/(www\.)?codepen\.io\/[\w-]+\/pen\/[\w-]+/],
-    oEmbedEndpoint: "https://codepen.io/api/oembed",
-    supportsCors: true,
-  },
-  {
-    name: "codesandbox",
-    patterns: [/^https?:\/\/(www\.)?codesandbox\.io\/s\/[\w-]+/],
-    oEmbedEndpoint: "https://codesandbox.io/oembed",
-    supportsCors: true,
-  },
-  {
-    name: "github-gist",
-    patterns: [/^https?:\/\/gist\.github\.com\/[\w-]+\/[\w]+/],
-    supportsCors: false,
-  },
-  // Design
-  {
-    name: "figma",
-    patterns: [
-      /^https?:\/\/(www\.)?figma\.com\/(file|design)\/[\w-]+/,
-      /^https?:\/\/(www\.)?figma\.com\/proto\/[\w-]+/,
-    ],
-    oEmbedEndpoint: "https://www.figma.com/api/oembed",
-    supportsCors: true,
-  },
-  {
-    name: "dribbble",
-    patterns: [/^https?:\/\/(www\.)?dribbble\.com\/shots\/[\w-]+/],
-    supportsCors: false,
-  },
-  // Music - Spotify and SoundCloud support CORS
-  {
-    name: "spotify",
-    patterns: [/^https?:\/\/open\.spotify\.com\/(track|album|playlist|episode|show)\/[\w]+/],
-    oEmbedEndpoint: "https://open.spotify.com/oembed",
-    supportsCors: true,
-  },
-  {
-    name: "soundcloud",
-    patterns: [/^https?:\/\/(www\.)?soundcloud\.com\/[\w-]+\/[\w-]+/],
-    oEmbedEndpoint: "https://soundcloud.com/oembed",
-    supportsCors: true,
-  },
-  // Other
-  {
-    name: "slideshare",
-    patterns: [/^https?:\/\/(www\.)?slideshare\.net\/[\w-]+\/[\w-]+/],
-    oEmbedEndpoint: "https://www.slideshare.net/api/oembed/2",
-    supportsCors: false,
-  },
-];
-
-/**
- * Detect provider from URL
- */
-export function detectVizelEmbedProvider(
-  url: string,
-  providers: readonly VizelEmbedProvider[] = vizelDefaultEmbedProviders
-): VizelEmbedProvider | null {
-  for (const provider of providers) {
-    for (const pattern of provider.patterns) {
-      if (pattern.test(url)) {
-        return provider;
-      }
-    }
-  }
-  return null;
-}
-
-/**
- * Check if URL is a valid embed URL
- */
-function isValidUrl(url: string): boolean {
-  try {
-    new URL(url);
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-/**
- * oEmbed response from provider (snake_case is standard oEmbed format)
- */
-interface OEmbedResponse {
-  type?: string;
-  html?: string;
-  width?: number;
-  height?: number;
-  title?: string;
-  description?: string;
-  // biome-ignore lint/style/useNamingConvention: oEmbed API uses snake_case
-  thumbnail_url?: string;
-  // biome-ignore lint/style/useNamingConvention: oEmbed API uses snake_case
-  provider_name?: string;
-  // biome-ignore lint/style/useNamingConvention: oEmbed API uses snake_case
-  author_name?: string;
-}
-
-/**
- * Fetch oEmbed data from a provider endpoint
- */
-async function fetchOEmbed(
-  url: string,
-  provider: VizelEmbedProvider
-): Promise<OEmbedResponse | null> {
-  if (!(provider.oEmbedEndpoint && provider.supportsCors)) {
-    return null;
-  }
-
-  try {
-    const oEmbedUrl = new URL(provider.oEmbedEndpoint);
-    oEmbedUrl.searchParams.set("url", url);
-    oEmbedUrl.searchParams.set("format", "json");
-
-    const response = await fetch(oEmbedUrl.toString(), {
-      headers: {
-        Accept: "application/json",
-      },
-    });
-
-    if (!response.ok) {
-      return null;
-    }
-
-    return (await response.json()) as OEmbedResponse;
-  } catch {
-    return null;
-  }
-}
-
-/**
- * Convert oEmbed response to EmbedData
- * Uses conditional spreading to avoid assigning undefined to optional properties
- */
-function oembedToEmbedData(
-  oembed: OEmbedResponse,
-  baseData: VizelEmbedData
-): VizelEmbedData | null {
-  // Check if we have HTML content (rich embed)
-  if (oembed.html) {
-    return {
-      ...baseData,
-      type: "oembed",
-      html: oembed.html,
-      ...(oembed.width !== undefined && { width: oembed.width }),
-      ...(oembed.height !== undefined && { height: oembed.height }),
-      ...(oembed.title !== undefined && { title: oembed.title }),
-      ...(oembed.description !== undefined && { description: oembed.description }),
-      ...(oembed.thumbnail_url !== undefined && { image: oembed.thumbnail_url }),
-      ...(oembed.provider_name !== undefined && { siteName: oembed.provider_name }),
-    };
-  }
-
-  // If we have title/thumbnail but no HTML, create OGP-like card
-  if (oembed.title || oembed.thumbnail_url) {
-    return {
-      ...baseData,
-      type: "ogp",
-      ...(oembed.title !== undefined && { title: oembed.title }),
-      ...(oembed.description !== undefined && { description: oembed.description }),
-      ...(oembed.thumbnail_url !== undefined && { image: oembed.thumbnail_url }),
-      ...(oembed.provider_name !== undefined && { siteName: oembed.provider_name }),
-    };
-  }
-
-  return null;
-}
-
-/**
- * Create default fetchEmbedData function that works client-side
- *
- * This function fetches oEmbed data for providers that support CORS.
- * For providers that don't support CORS or when oEmbed fails, it falls back to link type.
- *
- * For full OGP support, you need to provide a custom fetchEmbedData function
- * that fetches data through a server-side proxy.
- *
- * @param providers - Custom providers to use (defaults to vizelDefaultEmbedProviders)
- * @returns A fetchEmbedData function
- *
- * @example
- * ```typescript
- * const editor = useVizelEditor({
- *   features: {
- *     embed: {
- *       // Use default client-side fetcher
- *       fetchEmbedData: createVizelDefaultFetchEmbedData(),
- *     },
- *   },
- * });
- * ```
- */
-export function createVizelDefaultFetchEmbedData(
-  providers: readonly VizelEmbedProvider[] = vizelDefaultEmbedProviders
-): VizelFetchEmbedDataFn {
-  return async (url: string): Promise<VizelEmbedData> => {
-    const provider = detectVizelEmbedProvider(url, providers);
-
-    // Base embed data (conditionally include provider to avoid undefined)
-    const baseData: VizelEmbedData = {
-      url,
-      type: "link",
-      ...(provider?.name !== undefined && { provider: provider.name }),
-    };
-
-    // If no provider detected or doesn't support CORS, return as link
-    if (!(provider?.oEmbedEndpoint && provider.supportsCors)) {
-      return baseData;
-    }
-
-    // Try to fetch oEmbed
-    const oembedData = await fetchOEmbed(url, provider);
-    if (!oembedData) {
-      return baseData;
-    }
-
-    // Convert oEmbed response to EmbedData
-    const embedData = oembedToEmbedData(oembedData, baseData);
-    return embedData ?? baseData;
-  };
-}
+export {
+  createVizelDefaultFetchEmbedData,
+  detectVizelEmbedProvider,
+  type VizelEmbedData,
+  type VizelEmbedProvider,
+  type VizelEmbedType,
+  type VizelFetchEmbedDataFn,
+  vizelDefaultEmbedProviders,
+} from "../utils/embed-providers.ts";
+export { loadVizelEmbedScripts } from "./embed-renderers.ts";
 
 /**
  * Plugin key for embed paste handler
@@ -419,221 +101,16 @@ declare module "@tiptap/core" {
   }
 }
 
-// ============================================================================
-// NodeView Helper Functions
-// ============================================================================
-
 /**
- * Render loading state
+ * Check if URL is a valid embed URL
  */
-function renderLoading(container: HTMLElement): void {
-  const loader = document.createElement("div");
-  loader.className = "vizel-embed-loading";
-  loader.textContent = "Loading...";
-  container.appendChild(loader);
-}
-
-/**
- * Render oEmbed HTML content with sanitization.
- *
- * Uses DOMPurify to sanitize oEmbed HTML, allowing iframes (for
- * YouTube, Vimeo, etc.) while stripping scripts and event handlers.
- */
-function renderOembed(container: HTMLElement, html: string): void {
-  container.innerHTML = DOMPurify.sanitize(html, {
-    ADD_TAGS: ["iframe"],
-    ADD_ATTR: ["allow", "allowfullscreen", "frameborder", "scrolling"],
-  });
-}
-
-/**
- * Render OGP link preview card
- */
-function renderOgpCard(
-  container: HTMLElement,
-  attrs: {
-    url: string;
-    image?: string;
-    title?: string;
-    description?: string;
-    favicon?: string;
-    siteName?: string;
-  }
-): void {
-  const card = document.createElement("a");
-  card.href = attrs.url;
-  card.target = "_blank";
-  card.rel = "noopener noreferrer";
-  card.className = "vizel-embed-card";
-
-  if (attrs.image) {
-    const img = document.createElement("img");
-    img.src = attrs.image;
-    img.alt = attrs.title ?? "";
-    img.className = "vizel-embed-card-image";
-    card.appendChild(img);
-  }
-
-  const content = document.createElement("div");
-  content.className = "vizel-embed-card-content";
-
-  if (attrs.title) {
-    const title = document.createElement("div");
-    title.className = "vizel-embed-card-title";
-    title.textContent = attrs.title;
-    content.appendChild(title);
-  }
-
-  if (attrs.description) {
-    const desc = document.createElement("div");
-    desc.className = "vizel-embed-card-description";
-    desc.textContent = attrs.description;
-    content.appendChild(desc);
-  }
-
-  const meta = document.createElement("div");
-  meta.className = "vizel-embed-card-meta";
-
-  if (attrs.favicon) {
-    const favicon = document.createElement("img");
-    favicon.src = attrs.favicon;
-    favicon.alt = "";
-    favicon.className = "vizel-embed-card-favicon";
-    meta.appendChild(favicon);
-  }
-
-  if (attrs.siteName) {
-    const site = document.createElement("span");
-    site.textContent = attrs.siteName;
-    meta.appendChild(site);
-  }
-
-  if (meta.hasChildNodes()) {
-    content.appendChild(meta);
-  }
-
-  card.appendChild(content);
-  container.appendChild(card);
-}
-
-/**
- * Render simple link fallback
- */
-function renderLink(container: HTMLElement, url: string, title?: string): void {
-  const link = document.createElement("a");
-  link.href = url;
-  link.target = "_blank";
-  link.rel = "noopener noreferrer";
-  link.className = "vizel-embed-link";
-  link.textContent = title ?? url;
-  container.appendChild(link);
-}
-
-/**
- * Escape a value for inclusion inside a `vizel:` metadata comment.
- *
- * Order matters: `&` is escaped first so subsequent replacements do
- * not double-escape ampersand-introduced entities. The literal
- * `-->` sequence becomes `--&gt;` so the value cannot prematurely
- * close the host HTML comment.
- */
-function escapeMetadataCommentValue(value: string): string {
-  return value
-    .replace(/&/g, "&amp;")
-    .replace(/"/g, "&quot;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/-->/g, "--&gt;");
-}
-
-/**
- * Decode a value previously escaped by
- * {@link escapeMetadataCommentValue}.
- *
- * Reverses the same replacements in inverse order so the original
- * characters are restored verbatim on parse.
- */
-function unescapeMetadataCommentValue(value: string): string {
-  return value
-    .replace(/&gt;/g, ">")
-    .replace(/&lt;/g, "<")
-    .replace(/&quot;/g, '"')
-    .replace(/&amp;/g, "&");
-}
-
-/**
- * Parse the attribute list of a `<!-- vizel:embed key="value" -->`
- * comment into a plain object. Values are unescaped via
- * {@link unescapeMetadataCommentValue}.
- */
-function parseEmbedMetadataAttrs(attrsSrc: string): Record<string, string> {
-  const result: Record<string, string> = {};
-  const attrRegex = /([A-Za-z_][\w-]*)="([^"]*)"/g;
-  for (const match of attrsSrc.matchAll(attrRegex)) {
-    const key = match[1];
-    const value = match[2];
-    if (key === undefined || value === undefined) continue;
-    result[key] = unescapeMetadataCommentValue(value);
-  }
-  return result;
-}
-
-/**
- * Inline regex matching `[text](url)<!-- vizel:embed key="value" ... -->`.
- * Group 1 captures the link text, group 2 captures the URL, group 3
- * captures the comment body up to the terminating `-->`.
- */
-const VIZEL_EMBED_METADATA_PATTERN =
-  /^\[([^\]]*)\]\(([^)\s]+)\)<!--\s*vizel:embed\s+([\s\S]*?)\s*-->/;
-
-/**
- * Register a markdown-it inline rule that recognizes the
- * `[title](url)<!-- vizel:embed ... -->` pattern and emits a single
- * `vizel_embed` token carrying the recovered attributes. The rule
- * runs before `link` so the link parser never gets a chance to
- * consume the visible portion in isolation.
- */
-function registerEmbedRule(md: MarkdownIt): void {
-  md.inline.ruler.before("link", "vizel_embed", (state, silent) => {
-    if (state.src.charCodeAt(state.pos) !== 0x5b) return false;
-    const rest = state.src.slice(state.pos);
-    const match = VIZEL_EMBED_METADATA_PATTERN.exec(rest);
-    if (!match) return false;
-    const title = match[1] ?? "";
-    const url = match[2] ?? "";
-    const attrsSrc = match[3] ?? "";
-    if (!url) return false;
-    if (!silent) {
-      const attrs = parseEmbedMetadataAttrs(attrsSrc);
-      const token = state.push("vizel_embed", "div", 0);
-      token.content = title;
-      token.meta = { url, attrs };
-    }
-    state.pos += match[0].length;
+function isValidUrl(url: string): boolean {
+  try {
+    new URL(url);
     return true;
-  });
-
-  md.renderer.rules.vizel_embed = (tokens, idx) => {
-    const token = tokens[idx];
-    const meta = (token?.meta as { url?: string; attrs?: Record<string, string> } | undefined) ?? {
-      url: "",
-      attrs: {},
-    };
-    const url = meta.url ?? "";
-    const attrs = meta.attrs ?? {};
-    const title = token?.content ?? "";
-    const escapedUrl = url.replace(/&/g, "&amp;").replace(/"/g, "&quot;");
-    const titleAttr = attrs.title ?? title;
-    const escapedTitle = titleAttr.replace(/&/g, "&amp;").replace(/"/g, "&quot;");
-    const dataAttrs = Object.entries(attrs)
-      .filter(([key]) => key !== "title")
-      .map(
-        ([key, value]) =>
-          ` data-embed-${key}="${value.replace(/&/g, "&amp;").replace(/"/g, "&quot;")}"`
-      )
-      .join("");
-    return `<div data-vizel-embed="true" data-embed-url="${escapedUrl}" data-embed-title="${escapedTitle}"${dataAttrs}></div>`;
-  };
+  } catch {
+    return false;
+  }
 }
 
 /**
@@ -1125,33 +602,6 @@ export function createVizelEmbedExtension(options: VizelEmbedOptions = {}) {
 // =============================================================================
 // Embed View Helpers
 // =============================================================================
-
-/**
- * Re-parent any `<script>` tags inside an oEmbed container so the browser
- * actually executes them (innerHTML-inserted scripts are inert).
- *
- * Optionally invokes provider-specific bootstrap (Twitter widgets) when
- * the provider is `"twitter"`. Framework `VizelEmbedView` components call
- * this helper after their effect runs.
- */
-export function loadVizelEmbedScripts(container: HTMLElement, provider?: string): void {
-  const scripts = container.querySelectorAll("script");
-  for (const oldScript of scripts) {
-    const newScript = document.createElement("script");
-    for (const attr of Array.from(oldScript.attributes)) {
-      newScript.setAttribute(attr.name, attr.value);
-    }
-    if (oldScript.textContent) {
-      newScript.textContent = oldScript.textContent;
-    }
-    oldScript.parentNode?.replaceChild(newScript, oldScript);
-  }
-
-  if (provider === "twitter" && typeof window !== "undefined" && "twttr" in window) {
-    const twttr = (window as { twttr?: { widgets?: { load?: (el?: HTMLElement) => void } } }).twttr;
-    twttr?.widgets?.load?.(container);
-  }
-}
 
 /**
  * Discriminated view-model for the framework `VizelEmbedView` components.
